Split Init Pool into v2 and v3
