+++ src/entities/initPool/index.ts
@@ -1,40 +1,23 @@
-import { PoolType } from '../../types';
 import { InputValidator } from '../inputValidator/inputValidator';
 import { PoolState } from '../types';
-import { InitPoolComposableStable } from './composableStable/initPoolComposableStable';
-import {
-    InitPoolBase,
-    InitPoolBuildOutput,
-    InitPoolConfig,
-    InitPoolInput,
-} from './types';
-import { InitPoolWeighted } from './weighted/initPoolWeighted';
+import { InitPoolBuildOutput, InitPoolConfig, InitPoolInput } from './types';
+import { InitPoolV2 } from './initPoolV2';
+import { InitPoolV3 } from './initPoolV3';
 
 export * from './types';
 
 export class InitPool {
-    initPoolTypes: Record<string, InitPoolBase> = {};
-
     inputValidator: InputValidator = new InputValidator();
 
-    constructor(config?: InitPoolConfig) {
-        const { initPoolTypes: customAddLiquidityInitTypes } = config || {};
-        this.initPoolTypes = {
-            [PoolType.Weighted]: new InitPoolWeighted(),
-            [PoolType.ComposableStable]: new InitPoolComposableStable(),
-            ...customAddLiquidityInitTypes,
-        };
-    }
-
-    getInitPool(poolType: string): InitPoolBase {
-        if (!this.initPoolTypes[poolType]) {
-            throw new Error('Unsupported pool type: ${poolType}');
-        }
-        return this.initPoolTypes[poolType];
-    }
+    constructor(public config?: InitPoolConfig) {}
 
     buildCall(input: InitPoolInput, poolState: PoolState): InitPoolBuildOutput {
         this.inputValidator.validateAddLiquidity(input, poolState);
-        return this.getInitPool(poolState.type).buildCall(input, poolState);
+        switch (poolState.balancerVersion) {
+            case 2:
+                return new InitPoolV2().buildCall(input, poolState);
+            case 3:
+                return new InitPoolV3().buildCall(input, poolState);
+        }
     }
 }
